+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog for Laravel Repository
 
+## 2.0.0-dev
+
+- Suppression du système d'options (bancal).
+- Les fonctionnalités de filtrage de la méthode newQuery() ont été déportée dans
+  une nouvelle méthode : filter().
+- $order peut maintenant être fourni sous forme de tableau.
+
 ## 1.1.0 (2015-10-26)
 
 - Ajout de paramètres/fonctionnalités à la méthode newQuery(), ainsi qu'aux méthodes